Fix unconvert issues
$ gometalinter --vendor --deadline 10m --disable-all --enable=unconvert ./...
collector.go:218:10:warning: unnecessary conversion (unconvert)
collector.go:253:10:warning: unnecessary conversion (unconvert)
collector.go:296:61:warning: unnecessary conversion (unconvert)
collector.go:296:74:warning: unnecessary conversion (unconvert)
collector.go:311:19:warning: unnecessary conversion (unconvert)
collector.go:312:12:warning: unnecessary conversion (unconvert)
collector.go:313:12:warning: unnecessary conversion (unconvert)
collector.go:314:12:warning: unnecessary conversion (unconvert)
collector.go:315:12:warning: unnecessary conversion (unconvert)
collector.go:316:12:warning: unnecessary conversion (unconvert)
config/config.go:132:24:warning: unnecessary conversion (unconvert)
main.go:95:21:warning: unnecessary conversion (unconvert)

+++ main.go
@@ -92,7 +92,7 @@ func handler(w http.ResponseWriter, r *http.Request) {
 	// Delegate http serving to Prometheus client library, which will call collector.Collect.
 	h := promhttp.HandlerFor(registry, promhttp.HandlerOpts{})
 	h.ServeHTTP(w, r)
-	duration := float64(time.Since(start).Seconds())
+	duration := time.Since(start).Seconds()
 	snmpDuration.WithLabelValues(moduleName).Observe(duration)
 	log.Debugf("Scrape of target '%s' with module '%s' took %f seconds", target, moduleName, duration)
 }
